민주야 일해라 메시지 사진 보내기, 제안서 수정, 송금(고장)
메시지 사진 보내기, 제안서 수정, 송금(고장)

+++ src/main/java/com/hcn/hcn/model/WebUserData.java
@@ -31,9 +31,8 @@ public class WebUserData {
     private String websiteUrl; // 회사 사이트 URL
     @Column(columnDefinition = "TEXT")
     private String companyDescription; // 회사 소개글
-    private int points; // 보유 포인트
+    private float points; // 보유 포인트
 
-    // Getters and Setters
 
     @Override
     public String toString() {
